# openapi.yaml
openapi: 3.1.0
info:
  title: Archer API
  version: 0.1.0
  description: API for the Archer freelance marketplace. Amounts use integer minor units; Archer does not process payments.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register an account
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RegisterInput" } } } }
      responses: { "201": { description: Registered }, "409": { $ref: "#/components/responses/Conflict" }, "422": { $ref: "#/components/responses/ValidationError" } }
  /auth/login:
    post:
      tags: [Authentication]
      summary: Start a session
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LoginInput" } } } }
      responses: { "200": { description: Authenticated }, "401": { $ref: "#/components/responses/Unauthenticated" } }
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token
      responses: { "200": { description: Token rotated }, "401": { $ref: "#/components/responses/Unauthenticated" } }
  /auth/logout:
    post:
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      summary: Revoke the current session
      responses: { "204": { description: Logged out } }
  /auth/logout-all:
    post:
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      summary: Revoke every account session
      responses: { "204": { description: Logged out everywhere } }
  /auth/me:
    get:
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      summary: Get the current account and profiles
      responses: { "200": { description: Current account }, "401": { $ref: "#/components/responses/Unauthenticated" } }
...
  /contracts:
    get:
      tags: [Contracts]
      security: [{ bearerAuth: [] }]
      summary: List current user's contracts
      responses: { "200": { description: Contracts } }
  /contracts/{contractId}:
    get:
      tags: [Contracts]
      security: [{ bearerAuth: [] }]
      summary: Get a contract workspace
      parameters: [{ $ref: "#/components/parameters/ContractId" }]
      responses: { "200": { description: Contract } }
  /contracts/{contractId}/accept:
    post:
      tags: [Contracts]
      security: [{ bearerAuth: [] }]
      summary: Accept a contract offer
      parameters: [{ $ref: "#/components/parameters/ContractId" }]
      responses: { "200": { description: Active contract } }
  /contracts/{contractId}/complete:
    post:
      tags: [Contracts]
      security: [{ bearerAuth: [] }]
      summary: Request or finalize completion
      parameters: [{ $ref: "#/components/parameters/ContractId" }]
      responses: { "200": { description: Updated contract } }
  /milestones/{milestoneId}/submit:
    post:
      tags: [Contracts]
      security: [{ bearerAuth: [] }]
      summary: Submit milestone deliverables
      parameters: [{ $ref: "#/components/parameters/MilestoneId" }]
      responses: { "201": { description: Submission created } }
  /milestones/{milestoneId}/review:
    post:
      tags: [Contracts]
      security: [{ bearerAuth: [] }]
      summary: Approve a submission or request changes
      parameters: [{ $ref: "#/components/parameters/MilestoneId" }]
      responses: { "200": { description: Milestone updated } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    ContractId: { name: contractId, in: path, required: true, schema: { type: string } }
    MilestoneId: { name: milestoneId, in: path, required: true, schema: { type: string } }
...
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties: { email: { type: string, format: email }, password: { type: string, minLength: 8 } }
    RegisterInput:
      allOf:
        - { $ref: "#/components/schemas/LoginInput" }
        - type: object
          required: [displayName, modes]
          properties:
            displayName: { type: string, minLength: 2, maxLength: 80 }
            modes: { type: array, minItems: 1, maxItems: 2, items: { type: string, enum: [CLIENT, FREELANCER] } }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: { code: { type: string }, message: { type: string }, fields: { type: object, additionalProperties: { type: array, items: { type: string } } }, requestId: { type: string } }
  responses:
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Unauthenticated: { description: Authentication required, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    Conflict: { description: Resource conflict, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }